requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.9
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
filelock==3.16.1
identify==2.6.1
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.0.1
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.27.0
